refactor(proto)!: split the sandbox API along the control/data-plane seam
CORE-57. `sandbox.proto` was 634 lines behind eight section dividers,
which per the root CLAUDE.md rule is the signal to split — and here the
seam is architectural rather than cosmetic. It becomes the package
directory `arcbox/sandbox/v1/` (matching the package name, which is now
`arcbox.sandbox.v1`: the old flat `sandbox.v1` squatted a very generic
top-level name for a contract meant to be published):

  sandbox.proto    SandboxService            control plane
  process.proto    SandboxProcessService     data plane
  filesystem.proto SandboxFilesystemService  data plane
  snapshot.proto   SandboxSnapshotService    checkpoint / restore

Four services rather than one is the point: control-plane calls address
a fleet and can be served by a multi-tenant front door, while data-plane
calls carry a specific sandbox's stdio and file bytes and belong with
whatever is co-located with it. The daemon still serves all four on one
socket, so nothing changes locally.

The vsock-only port-forward payloads move to agent.proto, where the
other host↔guest wire messages live, with their own protocol enum — the
published contract must not be imported by the internal wire. Nothing
vsock-only ships in the sandbox package now.

prost emits one module per package, so all four files still generate a
single `arcbox.sandbox.v1.rs`, re-exported as `sandbox_v1`: consumer
imports are unchanged and the churn is confined to which client stub
each call site holds.
