+++ app/arcbox-api/src/lib.rs
@@ -22,9 +22,10 @@ pub use arcbox_grpc::v1::{
 #[cfg(target_os = "macos")]
 pub use arcbox_grpc::v1::{macos_service_client, macos_service_server};
 pub use arcbox_grpc::{
-    IconService, IconServiceServer, MigrationService, MigrationServiceServer, SandboxService,
-    SandboxServiceServer, SandboxSnapshotService, SandboxSnapshotServiceServer, SystemService,
-    SystemServiceServer,
+    IconService, IconServiceServer, MigrationService, MigrationServiceServer,
+    SandboxFilesystemService, SandboxFilesystemServiceServer, SandboxProcessService,
+    SandboxProcessServiceServer, SandboxService, SandboxServiceServer, SandboxSnapshotService,
+    SandboxSnapshotServiceServer, SystemService, SystemServiceServer,
 };
 
 pub use arcbox_protocol::v1::setup_status::Phase as SetupPhase;
@@ -32,8 +33,9 @@ pub use error::{ApiError, Result};
 #[cfg(target_os = "macos")]
 pub use grpc::MacosServiceImpl;
 pub use grpc::{
-    IconServiceImpl, KubernetesServiceImpl, MachineServiceImpl, SandboxServiceImpl,
-    SandboxSnapshotServiceImpl, SharedRuntime, StatsServiceImpl,
+    IconServiceImpl, KubernetesServiceImpl, MachineServiceImpl, SandboxFilesystemServiceImpl,
+    SandboxProcessServiceImpl, SandboxServiceImpl, SandboxSnapshotServiceImpl, SharedRuntime,
+    StatsServiceImpl,
 };
 pub use migration::MigrationServiceImpl;
 pub use system::{SetupState, SystemServiceImpl};
